refactor(presentation): move CarNameParser and RoundParser into parser subpackage
