+++ src/main/kotlin/racingcar/presentation/parser/CarNameParser.kt
@@ -1,4 +1,4 @@
-package racingcar.presentation
+package racingcar.presentation.parser
 
 object CarNameParser {
