+++ src/test/kotlin/racingcar/presentation/RoundParserTest.kt
@@ -3,6 +3,7 @@ package racingcar.presentation
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
+import racingcar.presentation.parser.RoundParser
 
 class RoundParserTest {
